Add timer to notification system
Nicks setup with a notify can now have an idle timer associated with
them. This option can be toggled off if the user does not want to
recieve idle highlight notices and only wants to recieve highligh
notices when they are not connected or when someone explicitly calls a
notify on them.

Default timeout is 900 seconds (15 minutes) and may be adjusted by each
user.