Sanitize woot title & grab just one price
Sanitize the Woot title of HTML entities. I'm aware that CGI::escapeHTML
probably isn't the best way but it works well enough. Also, We keep
getting a list of multiple prices for the price search. As the Item
we're handing back to people is just the first one we're going to clamp
the return to that one.